remove bree in favour of djs channel rename handling
