+++ package.json
@@ -24,7 +24,6 @@
     "@discordjs/builders": "^0.9.0",
     "@discordjs/rest": "^0.1.0-canary.0",
     "@prisma/client": "3.7.0",
-    "bree": "^7.1.4",
     "discord-api-types": "^0.25.2",
     "discord.js": "^13.3.1",
     "dotenv": "^10.0.0",
@@ -47,7 +46,6 @@
     "build": "yarn clean && yarn build:esbuild"
   },
   "devDependencies": {
-    "@types/bree": "^7.1.0",
     "@types/lodash": "^4.14.178",
     "@types/node": "^17.0.5",
     "@types/romans": "^2.0.0",
